refactor: Phase 2 PR5 — extract ColdStorageService read-side surface (#89)
Slice 2e of the Phase 2 structure refactor (#89), partial — read-side ops
extracted. Behavior byte-identical.

Stacked on PR4 #94 (SkipAnalysisService).

Extracted (src/main/services/cold-storage/):
- Storage-mode path helpers: getDownloadDir / getColdStorageDir / isAdvanced /
  dirsForScan. ~41 call sites in index.ts routed through coldStorageService.
- scanUsage (the StorageUsage scan: walks hot+cold roots, classifies bytes per
  anime/episode/bucket, emits STORAGE_USAGE_PROGRESS via injected broadcast).
- StorageUsage / StorageAnimeUsage / StorageEpisodeUsage / UsageEpisodeFiles
  types re-homed into the service module.

Deps-injected: store, downloadsFallbackDir (app.getPath('downloads')),
sanitizeFilename, parseEpisodeFromFilename, broadcast, usageProgressChannel.

Kept in index.ts (deferred — disk-mutation surface):
- deleteEpisodeFiles, episodeHasInProgressDownload, pruneDownloadedEpisode,
  episodeFileExists
- moveFileToCold, moveEpisodeToColdStorage, moveAllFilesToColdStorage
- CleanupCandidate, findCleanupCandidates, cleanupRunning, runWatchedCleanup

Issue #89 explicitly calls out manual spot-check for the move/cleanup ops;
isolating them in their own diff gives them a focused review path. They'll
follow as a stacked PR.

Tests: test/services/cold-storage.test.ts — 7 path-helper cases against
InMemoryStorage (simple/advanced mode, hot/downloadDir/fallback precedence,
dirsForScan composition). scanUsage is fs-heavy; covered by manual +
integration testing.

Full local quality gate green: typecheck, lint (0 errors), format:check,
41/41 vitest, build.
